+++ src/assets/src/js/filex-index.js
@@ -23,7 +23,7 @@ class filexIndex {
             keys.forEach((e, i, a) => {
                 params['uuids[' + i + ']'] = e;
             });
-            $("#filex-bulk-delete,#filex-bulk-acl,#filex-bulk-download").attr('data-params', JSON.stringify(params));
+            $("#filex-bulk-delete,#filex-bulk-acl,#filex-bulk-download").data('params', params);
         })
     }