+++ changelog.md
@@ -1,5 +1,7 @@
 # Changelog
-
+## 22-7-2020 0.6
+- [+] Add deep search function
+- [*] Improve inode query. Now permissions filter makes use of permissions in addition to roles
 ## 21-7-2020 0.5
 - [*] Fixed an issue in bulk download. Now downloaded files are the latest version.
 ## 16-6-2020 0.4